backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-environ==0.11.2
psycopg2-binary==2.9.9
django-filter==23.3
djangorestframework-simplejwt==5.3.0
Pillow==10.1.0
django-rest-swagger==2.2.0 